API client watch helper to retry disconnections from API server
