+++ manifests/namespace-install.yaml
@@ -360,6 +360,7 @@ spec:
         - /app
         - /shared
         image: argoproj/argocd-ui:latest
+        imagePullPolicy: Always
         name: ui
         volumeMounts:
         - mountPath: /shared
